requirements.txt
altair==6.0.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
courlan==1.3.2
dateparser==1.4.0
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.3
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
lxml==6.0.3
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tld==0.13.2
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
trafilatura==2.0.0
transformers==5.5.3
typer==0.24.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
